Add files via upload
A Password circuit breaker is an automatically operated electrical switch designed to protect an electrical circuit from damage caused by overload or short circuit. Its basic function is to detect a fault condition and interrupt current flow. Unlike a fuse, which operates once and then must be replaced, a circuit breaker can be reset (either manually or automatically) to resume normal operation. When operated manually fatal electrical accidents are seen to the line man are increasing during the electric line repair due to the lack of communication and coordination between the maintenance staff and the electric substation staff. In order to avoid such accidents, the breaker can be so designed such that only authorized person can operate it with a password.